Adapt to the use of BigDecimal
git-svn-id: https://scm.ops4j.org/repos/ops4j/projects/qi4j/apps/chronos@8218 9b982a3c-3ae5-0310-a4bc-d9a3335569bd

+++ core/src/main/java/org/qi4j/chronos/model/mixins/MoneyDescriptorMixin.java
@@ -12,6 +12,7 @@
  */
 package org.qi4j.chronos.model.mixins;
 
+import java.math.BigDecimal;
 import java.util.Currency;
 import org.qi4j.api.annotation.Uses;
 import org.qi4j.library.general.model.Descriptor;
@@ -33,7 +34,7 @@ public final class MoneyDescriptorMixin implements Descriptor
             currencyCode = Character.SPACE_SEPARATOR + currency.getCurrencyCode();
         }
 
-        Integer amount = money.getAmount();
+        BigDecimal amount = money.getAmount();
         return amount + currencyCode;
     }
